вынес классы в  модуль
