+++ HW10/Index.html
@@ -8,30 +8,30 @@
 </head>
 <body>
 <main class="main">
-<div class="task">
-  <h1> Завдання створити класи: </h1>
-  <ul>
-    <li>фігура: ( координати: х, у )</li>
-    <li> фігура -> круг ( координати: х, у ) -> радіус, площа(), периметр-длина окружности()</li>
-    <li> фігура: -> багатокутник, -> вершини[координати], периметр()</li>
-    <li> багатокутник:-> прямокутник -> довжина, висота, площа(), периметр()</li>
-    <li> прямокутник: -> квадрат -> довжинаСторони , площа(), периметр()</li>
-  </ul>
-  <p class="error"></p>
+  <div class="task">
+    <h1> Завдання створити класи: </h1>
+    <ul>
+      <li>фігура: ( координати: х, у )</li>
+      <li> фігура -> круг ( координати: х, у ) -> радіус, площа(), периметр-длина окружности()</li>
+      <li> фігура: -> багатокутник, -> вершини[координати], периметр()</li>
+      <li> багатокутник:-> прямокутник -> довжина, висота, площа(), периметр()</li>
+      <li> прямокутник: -> квадрат -> довжинаСторони , площа(), периметр()</li>
+    </ul>
+    <!--<p class="error"></p>-->
 
-  <!--<div class="form">-->
-    <!--<input class="txt js_x" placeholder=" Введите x">-->
-    <!--<input class="txt js_y" placeholder=" Введите У">-->
-    <!--<button class="btn js_bnt">Enter</button>-->
+    <!--<div class="form">-->
+      <!--<input class="txt js_x" placeholder=" Введите x">-->
+      <!--<input class="txt js_y" placeholder=" Введите У">-->
+      <!--<button class="btn js_bnt">Enter</button>-->
+    <!--</div>-->
+  </div>
+  <!--<div class="popup hide">-->
+    <!--<header class="popup_header">-->
+      <!--<span class="popup_name"> Результаты </span>-->
+      <!--<span class="close">X</span>-->
+    <!--</header>-->
   <!--</div>-->
-<!--</div>-->
-<!--<div class="popup hide">-->
-  <!--<header class="popup_header">-->
-    <!--<span class="popup_name"> результаты </span>-->
-    <!--<span class="close">X</span>-->
-  <!--</header>-->
-<!--</div>-->
 </main>
-  <script type="module" src="js/main.js"></script>
+<script type="module" src="js/main.js"></script>
 </body>
 </html>
