+++ HW10/css/style.css
@@ -20,51 +20,51 @@ body, html {
   box-shadow: 4px 2px 4px 2px #ccc;
 }
 
-/*.hide {*/
-  /*display: none;*/
-/*}*/
+.hide {
+  display: none;
+}
 
-/*.popup {*/
-  /*position: absolute;*/
-  /*left: 10%;*/
-  /*right: 10%;*/
-  /*width: 80%;*/
-  /*height: 500px;*/
-  /*border: 2px solid #5a6268;*/
-/*box-shadow: 4px 2px 4px 2px #ccc;*/
-/*}*/
+.popup {
+  position: absolute;
+  left: 10%;
+  right: 10%;
+  width: 80%;
+  height: 500px;
+  border: 2px solid #5a6268;
+box-shadow: 4px 2px 4px 2px #ccc;
+}
 
-/*.popup_header {*/
-  /*height: 35px;*/
-  /*width: 100%;*/
-  /*background: #ccc;*/
-  /*display: flex;*/
-  /*justify-content: space-between;*/
-  /*align-items: center;*/
-  /*color: #fff;*/
-/*}*/
+.popup_header {
+  height: 35px;
+  width: 100%;
+  background: #ccc;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  color: #fff;
+}
 
-/*.popup_name {*/
-  /*width: 765px;*/
-  /*padding-left: 25px;*/
-  /*text-transform: uppercase;*/
-/*}*/
+.popup_name {
+  width: 765px;
+  padding-left: 25px;
+  text-transform: uppercase;
+}
 
-/*.close {*/
-  /*line-height: 35px;*/
-  /*text-align: center;*/
-  /*top: 10px;*/
-  /*color: #ff0000;*/
-  /*width: 35px;*/
-  /*height: 32px;*/
-  /*border: 2px solid #ff0000;*/
-  /*cursor: pointer;*/
-/*}*/
-/*.close:hover {*/
-  /*transition: 0.5s;*/
-  /*color: #fff;*/
-  /*background: #ff0000;*/
-/*}*/
+.close {
+  line-height: 35px;
+  text-align: center;
+  top: 10px;
+  color: #ff0000;
+  width: 35px;
+  height: 32px;
+  border: 2px solid #ff0000;
+  cursor: pointer;
+}
+.close:hover {
+  transition: 0.5s;
+  color: #fff;
+  background: #ff0000;
+}
 
 
 h1 {
@@ -80,37 +80,37 @@ h1 {
   color: #dd0000;
 }
 
-/*.form {*/
-  /*height: 52px;*/
-  /*width: 600px;*/
-  /*padding: 0 100px;*/
-  /*margin: 30px auto;*/
-  /*display: flex;*/
-  /*justify-content: space-around;*/
-  /*align-items: center;*/
-  /*background: #ccc;*/
+.form {
+  height: 52px;
+  width: 600px;
+  padding: 0 100px;
+  margin: 30px auto;
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+  background: #ccc;
 
-/*}*/
+}
 
-/*.txt {*/
-  /*width: 150px;*/
-  /*height: 30px;*/
-  /*border: none;*/
-  /*border-radius: 10px;*/
-  /*padding-left: 15px;*/
-/*}*/
+.txt {
+  width: 150px;
+  height: 30px;
+  border: none;
+  border-radius: 10px;
+  padding-left: 15px;
+}
 
-/*.btn {*/
-  /*width: 100px;*/
-  /*height: 33px;*/
-  /*border: 2px solid #5a6268;*/
-  /*border-radius: 8px;*/
-  /*text-transform: uppercase;*/
-/*}*/
+.btn {
+  width: 100px;
+  height: 33px;
+  border: 2px solid #5a6268;
+  border-radius: 8px;
+  text-transform: uppercase;
+}
 
-/*.btn:hover {*/
-  /*transition: 0.5s;*/
-  /*color: #fff;*/
-  /*background: #5a6268;*/
-/*}*/
+.btn:hover {
+  transition: 0.5s;
+  color: #fff;
+  background: #5a6268;
+}
